comic_organizer/requirements.txt
requests
python-magic
rarfile
imagehash
guessit
Pillow
python-dotenv